+++ docs/src/examples/DC-DC converter.jl
@@ -47,7 +47,6 @@ const UT = DI.Utils
 const DO = DI.Domain
 const ST = DI.System
 const SY = DI.Symbolic
-const CO = DI.Control
 const OP = DI.Optim
 const AB = OP.Abstraction
 
@@ -83,7 +82,7 @@ concrete_controller = MOI.get(optimizer, MOI.RawOptimizerAttribute("concrete_con
 nstep = 300
 x0 = SVector(1.2, 5.6)
 x_traj, u_traj =
-    CO.get_closed_loop_trajectory(concrete_system.f, concrete_controller, x0, nstep)
+    ST.get_closed_loop_trajectory(concrete_system.f, concrete_controller, x0, nstep)
 
 fig = plot(; aspect_ratio = :equal);
 plot!(concrete_system.X);
