+++ problems/pwa_sys.jl
@@ -5,13 +5,11 @@ using MathematicalSystems, HybridSystems
 using FillArrays, CDDLib
 
 import Dionysos
-using Dionysos.Control
 using Dionysos.Problem
 const DI = Dionysos
 const UT = DI.Utils
 const DO = DI.Domain
 const ST = DI.System
-const CO = DI.Control
 const SY = DI.Symbolic
 const PR = DI.Problem
