+++ src/Dionysos.jl
@@ -4,7 +4,6 @@ include("utils/utils.jl")
 include("domain/domain.jl")
 include("system/system.jl")
 include("symbolic/symbolic.jl")
-include("control/control.jl")
 include("problem/problem.jl")
 include("optim/optim.jl")
 include("mapping/mapping.jl")
