+++ src/optim/abstraction/lazy_ellipsoids_abstraction.jl
@@ -11,7 +11,6 @@ const DI = Dionysos
 const UT = DI.Utils
 const DO = DI.Domain
 const ST = DI.System
-const CO = DI.Control
 const SY = DI.Symbolic
 const PR = DI.Problem
