+++ src/system/system.jl
@@ -9,6 +9,8 @@ using Plots
 
 using ..Utils
 UT = Utils
+
 include("controlsystem.jl")
 include("controller.jl")
+include("trajectory.jl")
 end
