+++ src/system/trajectory.jl
