+++ test/optim/test_ellipsoids_abstraction.jl
@@ -15,7 +15,6 @@ const UT = DI.Utils
 const DO = DI.Domain
 const ST = DI.System
 const SY = DI.Symbolic
-const CO = DI.Control
 const PR = DI.Problem
 const OP = DI.Optim
 const AB = OP.Abstraction
@@ -124,7 +123,7 @@ end
 
 # We simulate the closed loop trajectory
 x0 = concrete_problem.initial_set
-x_traj, u_traj, cost_traj = CO.get_closed_loop_trajectory(
+x_traj, u_traj, cost_traj = ST.get_closed_loop_trajectory(
     f_eval1,
     concrete_controller,
     cost_eval,
