+++ test/symbolic/test_symbolicmodel.jl
@@ -7,7 +7,6 @@ const DI = Dionysos
 const UT = DI.Utils
 const DO = DI.Domain
 const ST = DI.System
-const CO = DI.Control
 const SY = DI.Symbolic
 
 sleep(0.1) # used for good printing
